fixed a couple of GCC warning messages of -Wformat type, like "format %d expects argument of type 'int', but argument 2 has type 'size_t'"
Import internal commit 61ab75b04b0d9f7841aa433bd10858c74681ec62

+++ groups/bsl/bslstl/bslstl_multiset.t.cpp
@@ -4455,7 +4455,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase7_1()
             TestValues VALUES(SPEC);
 
             if (verbose) {
-                printf("\nFor an object of length %d:\n", LENGTH);
+                printf("\nFor an object of length %d:\n", (int)LENGTH);
                 P(SPEC);
             }
 
@@ -4573,7 +4573,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase7()
             const size_t      LENGTH = (int) strlen(SPEC);
 
             if (verbose) {
-                printf("\nFor an object of length %d:\n", LENGTH);
+                printf("\nFor an object of length %d:\n", (int)LENGTH);
                 P(SPEC);
             }
 
@@ -5213,7 +5213,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase3()
             Obj mX(&oa);
 
             if ((int)LENGTH != oldLen) {
-                if (verbose) printf("\tof length %d:\n", LENGTH);
+                if (verbose) printf("\tof length %d:\n", (int)LENGTH);
                  ASSERTV(LINE, oldLen <= (int)LENGTH);  // non-decreasing
                 oldLen = LENGTH;
             }
@@ -5409,7 +5409,8 @@ void TestDriver<KEY, COMP, ALLOC>::testCase2()
 
             if (0 < LENGTH) {
                 if (verbose) {
-                    printf("\t\tOn an object of initial length %d.\n", LENGTH);
+                    printf("\t\tOn an object of initial length %d.\n",
+                           (int)LENGTH);
                 }
 
                 for (size_t tj = 0; tj < LENGTH - 1; ++tj) {
